requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
Click==7.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mpmath==1.1.0
pathspec==0.7.0
pylint==2.4.4
regex==2020.1.8
six==1.14.0
sympy==1.5.1
toml==0.10.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
